Remove firabase dependency on Charts
